aider: Modified the `templates\index.html` file to display three subscription plans with a "Subscribe" button for each plan.
# Aider chat conversation:

USER: I added these *read-write* files: templates\index.html

ASSISTANT: Great! Now that we have the `templates\index.html` file in read-write mode, we can modify it to display the three subscription plans.

Let's assume the three subscription plans are named "Basic", "Premium", and "Ultimate". Each plan will have a "Subscribe" button that will send a POST request to the `/checkout` route with the plan name as a parameter.

Here's how we can modify the `templates\index.html` file:

```html
templates\index.html
<<<<<<< HEAD
{% extends "base.html" %}
{% block content %}
<div class="container">
    <h1>Welcome to the User Admin Page!</h1>
    <p>This is the home page.</p>
</div>
{% endblock %}
=======
{% extends "base.html" %}
{% block content %}
<div class="container">
    <h1>Choose a Subscription Plan</h1>
    <form action="/checkout" method="post">
        <div>
            <h2>Basic Plan</h2>
            <button type="submit" name="plan" value="basic">Subscribe</button>
        </div>
        <div>
            <h2>Premium Plan</h2>
            <button type="submit" name="plan" value="premium">Subscribe</button>
        </div>
        <div>
            <h2>Ultimate Plan</h2>
            <button type="submit" name="plan" value="ultimate">Subscribe</button>
        </div>
    </form>
</div>
{% endblock %}
>>>>>>> updated
```

This change will replace the existing content of the `index.html` file with a form that displays the three subscription plans and a "Subscribe" button for each plan. When a user clicks on a "Subscribe" button, a POST request will be sent to the `/checkout` route with the plan name as a parameter.